fix file folder structure
