+++ app/components/NotificationContainer.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script setup lang="ts">
-import { useNotification } from '~/composables/useNotification'
+import { useNotification } from '../composables/useNotification'
 
 const { notifications, remove } = useNotification()
 </script>
