+++ app/composables/useNotification.ts
