+++ app/middleware/auth.ts
