+++ app/plugins/auth.client.ts
@@ -1,5 +1,5 @@
 export default defineNuxtPlugin(async () => {
-  const { useAuthStore } = await import('~/stores/auth')
+  const { useAuthStore } = await import('../../stores/auth')
   const authStore = useAuthStore()
   
   // Initialize auth state from localStorage
