+++ app/plugins/auto-logout.client.ts
@@ -1,3 +1,8 @@
+import { useAutoLogout } from '../composables/useAutoLogout'
+import { useAuth } from '../composables/useAuth'
+import { useRouter } from 'vue-router'
+import { watch } from 'vue'
+
 export default defineNuxtPlugin(() => {
   const { startAutoLogout, stopAutoLogout } = useAutoLogout()
   const { isAuthenticated, logout } = useAuth()
